+++ src/fetcher.ts
@@ -29,47 +29,28 @@ const defaultFetch: typeof window.fetch = async (resource, init) => {
   if (typeof window === "object" && typeof require !== "function") {
     return window.fetch(resource, init);
   }
-  /* istanbul ignore if: `require` is always defined in the unit test environment */
-  if (typeof require !== "function") {
-    // When using Node.js with ES Modules, require is not defined:
-    let fetch;
-    try {
-      // solid-client-authn-browser may be unresolved, we just try to autodetect it.
-      const { fetch: defaultSessionFetch } = await import(
-        /* eslint-disable import/no-unresolved */
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        "@inrupt/solid-client-authn-browser"
-      );
-      fetch = defaultSessionFetch;
-    } catch (e) {
-      const crossFetchModule = await import("cross-fetch");
-      fetch = crossFetchModule.default;
-      return await fetch(resource, init);
-    }
-  }
-  // Implementation note: it's up to the client application to resolve these module names to their
-  // respective npm packages. At least one commonly used tool (Webpack) is only able to do that if
-  // the module names are literal strings.
-  // Additionally, Webpack throws a warning in a way that halts compilation for at least Next.js
-  // when using native JavaScript dynamic imports (`import()`), whereas `require()` just logs a
-  // warning. Since the use of package names instead of file names requires a bundler anyway, this
-  // should not have any practical consequences. For more background, see:
-  // https://github.com/webpack/webpack/issues/7713
+
   // eslint-disable-next-line no-shadow
-  let fetch;
+  // let fetch;
+  // try {
+  //   // solid-client-authn-browser may be unresolved, we just try to autodetect it.
+  //   const { fetch: defaultSessionFetch } = await import(
+  //     /* eslint-disable import/no-unresolved */
+  //     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  //     // @ts-ignore
+  //     "@inrupt/solid-client-authn-browser"
+  //   );
+  //   /* istanbul ignore next : `solid-client-authn-browser` is not a dependency of this library */
+  //   fetch = defaultSessionFetch;
+  // } catch (e) {
+  //   const crossFetchModule = await import("cross-fetch");
+  //   fetch = crossFetchModule.default;
+  //   // return await fetch(resource, init);
+  // }
+  // return fetch(resource, init);
 
-  try {
-    // eslint-disable-next-line global-require, @typescript-eslint/no-var-requires
-    const sca = require("@inrupt/solid-client-authn-browser");
-    /* istanbul ignore next : `solid-client-authn-browser` is not a dependency of this library */
-    fetch = sca.fetch;
-  } catch (e) {
-    // eslint-disable-next-line prefer-const, global-require
-    fetch = require("cross-fetch");
-  }
-
-  return fetch(resource, init);
+  const crossFetchModule = await import("cross-fetch");
+  return crossFetchModule.default(resource, init);
 };
 
 export default defaultFetch;